Add files via upload
There seems to be an issue with 4k screens. I dont really know what is causing it but i changed some functions in the script to use the old calculation method for height. I hope this fixes it, it does work on wider and normal 1080p screens but since i dont own a 4k display i cant test it.